requirements.txt
python-telegram-bot==21.7
notion-client==2.2.1
python-dotenv==1.0.0
mistralai==1.2.2
